requirements.txt
numpy 
pandas
python-dotenv
mysql-connector-python
PyMySQL
scikit-learn
# -e .